safari-600.7-branch/LayoutTests/js/dfg-cse-dead-get-scoped-var-expected.txt

Tests that CSE doesn't try to match against a dead GetScopedVar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(i)(false) is 42 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
